+++ colors/operation_color.c
@@ -6,7 +6,7 @@
 /*   By: atucci <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/24 18:28:35 by atucci            #+#    #+#             */
-/*   Updated: 2024/07/30 12:37:03 by atucci           ###   ########.fr       */
+/*   Updated: 2024/07/30 13:17:48 by atucci           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,9 +39,9 @@ t_color	multiply_color_by_scalar(t_color c, float scalar)
 {
 	t_color	result;
 
-	result.r = my_clamp((int)(c.r * scalar), 0, 255);
-	result.g = my_clamp((int)(c.g * scalar), 0, 255);
-	result.b = my_clamp((int)(c.b * scalar), 0, 255);
+	result.r = my_clamp((c.r * scalar), 0, 255);
+	result.g = my_clamp((c.g * scalar), 0, 255);
+	result.b = my_clamp((c.b * scalar), 0, 255);
 	return (result);
 }
 
@@ -50,13 +50,13 @@ t_color	multiply_colors(t_color c1, t_color c2)
 {
 	t_color	result;
 
-	result.r = my_clamp((int)((c1.r / 255.0) * (c2.r / 255.0) * 255), 0, 255);
-	result.g = my_clamp((int)((c1.g / 255.0) * (c2.g / 255.0) * 255), 0, 255);
-	result.b = my_clamp((int)((c1.b / 255.0) * (c2.b / 255.0) * 255), 0, 255);
+	result.r = my_clamp(((c1.r / 255.0) * (c2.r / 255.0) * 255), 0, 255);
+	result.g = my_clamp(((c1.g / 255.0) * (c2.g / 255.0) * 255), 0, 255);
+	result.b = my_clamp(((c1.b / 255.0) * (c2.b / 255.0) * 255), 0, 255);
 	return (result);
 }
 
-t_color	create_color(int r, int g, int b)
+t_color	create_color(double r, double g, double b)
 {
 	t_color	new;
 
@@ -66,6 +66,7 @@ t_color	create_color(int r, int g, int b)
 	return (new);
 }
 
+/*
 int main()
 {
 	printf("%sSCENARIO: Adding Colors%s\n", RED, RESET);
@@ -102,4 +103,4 @@ int main()
 
 	return 0;
 }
-
+*/
